Main file created; relevant libraries found.
After encountering significant difficulty with dependency and version errors when trying to implement a .json parsing library (https://github.com/open-source-parsers/jsoncpp) and a GUI library called SFML (https://www.sfml-dev.org/index.php) for C++, the decision was made to switch our main language for this project to Python.

This is beneficial, as built-in libraries exist for GUI development (tkinter) and json parsing (json). A main Python file was created with comments for intended functionality, to be implemented soon by the team.